Add two faster implementations of Levenshtein distance.
The Lev1.hs and Lev2.hs modules implement a Levenshtein distance
function that iteratively computes sub-distances one diagonal
at a time in the (n+1) * (m+1) matrix L where

 L[i,j] = Levenshtein distance for
             (drop i as)
             (drop j bs)

The lower-right corner is L[n,m] = 0(if indexing is zero based),
the Levenshtein distance between two empty lists. The next diagonal
in is

  L[n-1,m] = 1
  L[n,m-1] = 1

which are the distances between empty lists and singleton lists.
With those two diagonals, we can compute the next diagonal, and
so on, until we reach L[0,0], which is the desired Levenshtein
distance.

In fact, these functions are written so they use an empty diagonal
beyond L[n,m]; the diagonal is empty because it's intersection
with the (n+1)*(m+1) grid is empty.

The functions share a Grid module that translates two-dimensional
matrix coordinates into an index into a diagonal, where zero
is the location of the leftmost and lowest element in the diagonal.

The difference between Lev1 and Lev2 is small but significant.
When computing a single Levenshtein distance L[i,j] from the
three distances

  del   = L[i+1,j]
  ins   = L[i,j+1]
  subst = L[i+1,j+1]

the Lev1 version does this as

  minimum $ catMaybes $ [del, ins, subst]

The 'Maybes' arise because we achieve uniformity by allowing references
to non-existent matric positions like L[n+1,m+1]. For non-existent
positions we compute Nothing, while for legitimate positions we
wrap the distance with 'Just'.

In Lev1, we therefore create a list of 'Maybe Int', compute a list of
'Int' from that, and then take the minimum. Profiling suggested
that the time and allocation cost was significant, so in Lev2
I instead defined a function

 mm mx my = maybe my (\x -> Just $ maybe x (min x) my) mx

which computes the minimum of two 'Maybe Int' values. The result
is Nothing if and only if both values are Nothing.

Run times with 'time', profiling, and Criterion bench-marking
all show that there is a 30 percent to 50 percent drop in the
runtime using the Lev2 implementation.

There is still a lot of improvement that I'm sure could be made
here. In particular:

1. use the ST monad and mutable arrays for the diagonals.
   With three arrays of length (min n m) we can allocate
   the arrays once and then rotate through them, instead
   of creating new vectors on each iteration.

2. don't use two-dimensional coordinates at all. Instead,
   figure out how to rework the computations using only
   diagonal coordinates. This will allow us to avoid creating
   ordered pairs and lists of ordered pairs, cutting down
   even further on memory allocation and the time to build
   pairs and to access their components.

I've added tests for the Lev1 and Lev2 implementations, a
Criterion benchmark (levn-bench.hs), and two tiny programs
to use GHC profiling on (lev1-prof.hs, lev2-prof.hs).