Add the Lev5 implementation, the best one so far.
I realized I was being a dumb-ass by computing the elements of
the Levenshtein distance matrix a diagnonal at a time. It can
also be computed a row at a time. This requires less memory
overall and is also considerably simpler. Best of all, this
implementation is a *lot* faster than the others. Even the
unit tests on small inputs run noticeably faster.

For example, on Benny's benchmark comparing "kitten"*300
with "sitting"*300, this computes the result in about 570
milliseconds on my machine, compared to about 4.75 seconds
for the Lev2 implementation.

I think I'll experiment with a Lev6 that is the same
idea done with mutable vectors in the ST monad. I'd like
to see if there is the same horrendous slowdown that
I saw with Lev2 versus Lev4.